Conflicts:
	duel-quiz-server/src/duel/quiz/server/controller/QuestionController.java

+++ duel-quiz-client/src/duel/quiz/client/model/Duel.java
@@ -9,7 +9,7 @@ package duel.quiz.client.model;
  * @author corteshs
  */
 public class Duel {
-    private long duelID;
+    private int duelID;
     //TODO Enum for String
     private String status;
     //How to determine which one is which
@@ -56,11 +56,11 @@ public class Duel {
     
     
 
-    public long getDuelID() {
+    public int getDuelID() {
         return duelID;
     }
 
-    public void setDuelID(long duelID) {
+    public void setDuelID(int duelID) {
         this.duelID = duelID;
     }
 
@@ -88,12 +88,12 @@ public class Duel {
         this.scorePlayer2 = scorePlayer2;
     }
 
-    public Duel(long duelID, String status) {
+    public Duel(int duelID, String status) {
         this.duelID = duelID;
         this.status = status;
     }
 
-    public Duel(long duelID, String status, int scorePlayer1, int scorePlayer2) {
+    public Duel(int duelID, String status, int scorePlayer1, int scorePlayer2) {
         this.duelID = duelID;
         this.status = status;
         this.scorePlayer1 = scorePlayer1;
